+++ bot/events/connected.js
@@ -95,8 +95,15 @@ module.exports = function(bot, db) {
 
       pointReset(bot, db);
 
-      var complete = (Date.now() - initStart) / 1000;
-      bot.sendChat(`\`Initialization completed in ${complete} seconds\``);
+      bot.sendChat(`\`Initialization complete\``);
+
+      Promise.all([
+        dmStore.init(bot),
+        historyStore.init(bot)
+      ]).then(function(){
+        var complete = (Date.now() - initStart) / 1000;
+        bot.log('info', 'BOT', `Initialization completed in ${complete} seconds`);
+      });
     }, 3000);
   });
 };